// src/front/src/api/gateway/endpoints/auth/auth.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Gateway API
 * OpenAPI spec version: v1
 */
import { useMutation } from "@tanstack/react-query";
import type {
	MutationFunction,
	QueryClient,
	UseMutationOptions,
	UseMutationResult,
} from "@tanstack/react-query";

import type {
	ApiErrorResponse,
	LoginRequest,
	LoginResponse,
} from "../../models";

import { customFetch } from "../../../custom-fetch";

type SecondParameter<T extends (...args: never) => unknown> = Parameters<T>[1];

export type loginResponse200 = {
	data: LoginResponse;
	status: 200;
};

export type loginResponse400 = {
	data: ApiErrorResponse;
	status: 400;
};

export type loginResponseSuccess = loginResponse200 & {
	headers: Headers;
};
export type loginResponseError = loginResponse400 & {
	headers: Headers;
};

export type loginResponse = loginResponseSuccess | loginResponseError;

export const getLoginUrl = () => {
	return `/api/auth/login`;
};

export const login = async (
	loginRequest: LoginRequest,
	options?: RequestInit,
): Promise<loginResponse> => {
	return customFetch<loginResponse>(getLoginUrl(), {
		...options,
		method: "POST",
		headers: { "Content-Type": "application/json", ...options?.headers },
		body: JSON.stringify(loginRequest),
	});
};

export const getLoginMutationOptions = <
	TError = ApiErrorResponse,
	TContext = unknown,
>(options?: {
	mutation?: UseMutationOptions<
		Awaited<ReturnType<typeof login>>,
		TError,
		{ data: LoginRequest },
		TContext
	>;
	request?: SecondParameter<typeof customFetch>;
}): UseMutationOptions<
	Awaited<ReturnType<typeof login>>,
	TError,
	{ data: LoginRequest },
	TContext
> => {
	const mutationKey = ["login"];
	const { mutation: mutationOptions, request: requestOptions } = options
		? options.mutation &&
			"mutationKey" in options.mutation &&
			options.mutation.mutationKey
			? options
			: { ...options, mutation: { ...options.mutation, mutationKey } }
		: { mutation: { mutationKey }, request: undefined };

	const mutationFn: MutationFunction<
		Awaited<ReturnType<typeof login>>,
		{ data: LoginRequest }
	> = (props) => {
		const { data } = props ?? {};

		return login(data, requestOptions);
	};

	return { mutationFn, ...mutationOptions };
};

export type LoginMutationResult = NonNullable<
	Awaited<ReturnType<typeof login>>
>;
export type LoginMutationBody = LoginRequest;
export type LoginMutationError = ApiErrorResponse;

export const useLogin = <TError = ApiErrorResponse, TContext = unknown>(
	options?: {
		mutation?: UseMutationOptions<
			Awaited<ReturnType<typeof login>>,
			TError,
			{ data: LoginRequest },
			TContext
		>;
		request?: SecondParameter<typeof customFetch>;
	},
	queryClient?: QueryClient,
): UseMutationResult<
	Awaited<ReturnType<typeof login>>,
	TError,
	{ data: LoginRequest },
	TContext
> => {
	const mutationOptions = getLoginMutationOptions(options);

	return useMutation(mutationOptions, queryClient);
};